Extract error list component
