+++ apps/client-react/src/components/error-list.module.css
@@ -1,0 +1,9 @@
+.errorList {
+	margin: 0;
+	padding: 0;
+	list-style-type: none;
+
+	li {
+		color: var(--color-danger);
+	}
+}
