+++ apps/client-react/src/views/project-create-route.module.css
@@ -98,13 +98,3 @@
 	padding: 0.25rem 0.25rem;
 	font-size: 0.9rem;
 }
-
-.errorList {
-	margin: 0;
-	padding: 0;
-	list-style-type: none;
-
-	li {
-		color: var(--color-danger);
-	}
-}
